Fix minor Bug, US/ADC/readMotor Task added

+++ include/portmanager.h
@@ -206,7 +206,7 @@ public:
 
 	bool digitalRead();
 
-	float analogRead(_ADMUX vRef = AVCC, _ADCSRA_PRESCALER prescaler = F_CPU_BY_128, _ADCSRB_AUTOTRIGGER autoTrigger = FREE_RUNNING_MODE);
+	uint16_t analogRead(_ADMUX vRef = AVCC, _ADCSRA_PRESCALER prescaler = F_CPU_BY_128, _ADCSRB_AUTOTRIGGER autoTrigger = FREE_RUNNING_MODE);
 	void getPinData();
 
 	uint8_t getPinNumber();
